feat:add modules solution
